Refactor provider as hook
